Create Dice Solitaire
As their turn begin, a player has 0 turn-points. They can either roll the dice or hold. When they roll there are two possible outcomes. Most of the time the sum of the dice is added to the person’s turn-points and they get another turn. But if the sum is 2, 7 or 12 then they Bust, they lose all turn-points,and their turn is over. Once they get to 50 points, they win.